Extract get-enhanced-consent to its own module and types
