--- static/src/javascripts/projects/commercial/types.ts
@@ -1,11 +1,0 @@
-import type {
-	ConsentState,
-	Framework,
-} from '@guardian/consent-management-platform/dist/types';
-
-type ConsentStateEnhanced = ConsentState & {
-	canTarget: boolean;
-	framework: Framework | null;
-};
-
-export { ConsentStateEnhanced };
