+++ static/src/javascripts/projects/common/modules/commercial/observable-store.ts
@@ -1,67 +1,9 @@
 import { onConsentChange } from '@guardian/consent-management-platform';
-import type {
-	ConsentState,
-	Framework,
-} from '@guardian/consent-management-platform/dist/types';
-import type { TCFv2ConsentList } from '@guardian/consent-management-platform/dist/types/tcfv2';
+import type { ConsentState } from '@guardian/consent-management-platform/dist/types';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import type { ConsentStateEnhanced } from 'commercial/types';
 import { getPageTargeting } from './build-page-targeting';
-
-const tcfv2ConsentedToAll = (consents: TCFv2ConsentList): boolean => {
-	return (
-		Object.keys(consents).length > 0 &&
-		Object.values(consents).every(Boolean)
-	);
-};
-
-const buildConsentStateEnhanced = (
-	consentState: ConsentState,
-	canTarget: boolean,
-	framework: Framework | null,
-): ConsentStateEnhanced => ({
-	...consentState,
-	canTarget,
-	framework,
-});
-
-const enhanceConsentState = (state: ConsentState): ConsentStateEnhanced => {
-	if (state.tcfv2) {
-		return buildConsentStateEnhanced(
-			state,
-			tcfv2ConsentedToAll(state.tcfv2.consents),
-			'tcfv2',
-		);
-	} else if (state.ccpa) {
-		return buildConsentStateEnhanced(state, !state.ccpa.doNotSell, 'ccpa');
-	} else if (state.aus) {
-		return buildConsentStateEnhanced(
-			state,
-			state.aus.personalisedAdvertising,
-			'aus',
-		);
-	}
-	return buildConsentStateEnhanced(state, false, null);
-};
-
-const getEnhancedConsent = (): Promise<ConsentStateEnhanced> =>
-	new Promise<ConsentStateEnhanced>((resolve, reject) => {
-		onConsentChange((consentState) => {
-			if (consentState.tcfv2) {
-				// For tcfv2 only, the first onConsentChange is fired before the user has
-				// interacted with the consent banner. We want to ignore this first consent state.
-				if (consentState.tcfv2.eventStatus !== 'cmpuishown') {
-					resolve(enhanceConsentState(consentState));
-				}
-				return;
-			} else if (consentState.ccpa || consentState.aus) {
-				resolve(enhanceConsentState(consentState));
-			}
-
-			reject('Unknown framework');
-		});
-	});
+import { enhanceConsentState } from './get-enhanced-consent';
 
 const consentObservable = new Observable<ConsentState>((subscriber) => {
 	onConsentChange((consentState) => {
@@ -81,4 +23,4 @@ const pageTargetingObservable = consentObservable.pipe(
 	map((consentState) => [consentState, getPageTargeting(consentState)]),
 );
 
-export { consentObservable, getEnhancedConsent, pageTargetingObservable };
+export { consentObservable, pageTargetingObservable };
